+++ app/files/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+files_bp = Blueprint("files", __name__, template_folder="templates")
+
+from . import routes