+++ app/helpers/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+helpers_bp = Blueprint("helpers", __name__, template_folder="templates")
+
+from . import routes