+++ tests/test_helpers/test_helpers_bp.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+
+
+def test_import_blueprint():
+    from app.helpers import helpers_bp
+
+    assert isinstance(helpers_bp, Blueprint)