+++ tests/test_utils.py
@@ -1,4 +1,5 @@
 import json
+import sys
 import warnings
 from datetime import datetime
 from unittest import mock
@@ -6,17 +7,17 @@ from unittest import mock
 import pytest
 
 from app.utils import (
-    HidesWarning,
-    SudoersWarning,
-    add_to_hides,
+    add_to_ignored,
     gen_random_password,
     get_folders,
-    get_hides,
+    get_ignored,
+    get_post_arg,
     get_sudoers,
     get_user,
     log,
-    remove_from_hides,
+    remove_from_ignored,
 )
+from app.utils.exceptions import IngoredWarning, SudoersWarning
 
 
 def test_warning():
@@ -39,62 +40,63 @@ def test_get_sudoers(mocker):
         assert get_sudoers() == []
 
 
-@mock.patch("app.utils.cfg.HIDE_PATH")
-class TestGetHides:
-    def test_normal(self, hide_path_mock):
+@mock.patch("app.utils.cfg.IGNORED_PATH")
+class TestGetignored:
+    def test_normal(self, ignored_path_mock):
         data = ["bbb", "aaa"]
-        hide_path_mock.read_text.return_value = json.dumps(data * 5)
+        ignored_path_mock.read_text.return_value = json.dumps(data * 5)
 
-        hides = get_hides()
+        ignored = get_ignored()
         data.sort()
-        assert hides == data
+        assert ignored == data
 
-    def test_error(self, hide_path_mock):
-        hide_path_mock.read_text.return_value = ""
+    def test_error(self, ignored_path_mock):
+        ignored_path_mock.read_text.return_value = ""
 
-        with pytest.warns(HidesWarning):
-            hides = get_hides()
-        assert hides == []
+        with pytest.warns(IngoredWarning):
+            ignored = get_ignored()
+        assert ignored == []
 
 
-@mock.patch("app.utils.cfg.HIDE_PATH")
-@mock.patch("app.utils.get_hides")
-class TestAddToHides:
-    def test_true(self, hides_mock, hide_path_mock):
-        hides_mock.return_value = ["aaa", "bbb"]
-        result = add_to_hides("ccc")
+@mock.patch("app.utils.cfg.IGNORED_PATH")
+@mock.patch("app.utils.get_ignored")
+class TestAddToignored:
+    def test_true(self, ignored_mock, ignored_path_mock):
+        ignored_mock.return_value = ["aaa", "bbb"]
+        result = add_to_ignored("ccc")
 
         expected_call = json.dumps(["aaa", "bbb", "ccc"], indent=4)
         assert result is True
-        hide_path_mock.write_text.assert_called_once_with(expected_call)
+        ignored_path_mock.write_text.assert_called_once_with(expected_call)
 
-    def test_false(self, hides_mock, hide_path_mock):
-        hides_mock.return_value = ["aaa", "bbb", "ccc"]
-        result = add_to_hides("ccc")
+    def test_false(self, ignored_mock, ignored_path_mock):
+        ignored_mock.return_value = ["aaa", "bbb", "ccc"]
+        result = add_to_ignored("ccc")
 
         assert result is False
-        hide_path_mock.write_text.assert_not_called()
+        ignored_path_mock.write_text.assert_not_called()
 
 
-@mock.patch("app.utils.cfg.HIDE_PATH")
-@mock.patch("app.utils.get_hides")
-class TestRemoveFromHides:
-    def test_true(self, hides_mock, hide_path_mock):
-        hides_mock.return_value = ["aaa", "bbb", "ccc"]
-        result = remove_from_hides("ccc")
+@mock.patch("app.utils.cfg.IGNORED_PATH")
+@mock.patch("app.utils.get_ignored")
+class TestRemoveFromignored:
+    def test_true(self, ignored_mock, ignored_path_mock):
+        ignored_mock.return_value = ["aaa", "bbb", "ccc"]
+        result = remove_from_ignored("ccc")
 
         expected_call = json.dumps(["aaa", "bbb"], indent=4)
         assert result is True
-        hide_path_mock.write_text.assert_called_once_with(expected_call)
+        ignored_path_mock.write_text.assert_called_once_with(expected_call)
 
-    def test_false(self, hides_mock, hide_path_mock):
-        hides_mock.return_value = ["aaa", "bbb"]
-        result = remove_from_hides("ccd")
+    def test_false(self, ignored_mock, ignored_path_mock):
+        ignored_mock.return_value = ["aaa", "bbb"]
+        result = remove_from_ignored("ccd")
 
         assert result is False
-        hide_path_mock.write_text.assert_not_called()
+        ignored_path_mock.write_text.assert_not_called()
 
 
+@pytest.mark.skipif(sys.platform != "win32", reason="Does not work on linux")
 @mock.patch("app.utils.cfg", spec=True)
 @mock.patch("app.utils.request", spec=True)
 @mock.patch("app.utils.asctime", spec=True)
@@ -118,6 +120,7 @@ def test_log(asctime_mock, request_mock, cfg_mock):
     assert fp.write.call_count == 2
 
 
+@pytest.mark.skipif(sys.platform != "win32", reason="Does not work on linux")
 @mock.patch("app.utils.request", spec=True)
 class TestGetUser:
     def test_no_authorization(self, request_mock):
@@ -162,20 +165,26 @@ class TestGetFolders:
         os_walk_mock = mock.patch("os.walk", spec=True).start()
         get_user_mock = mock.patch("app.utils.get_user", spec=True).start()
         sudoers_mock = mock.patch("app.utils.get_sudoers", spec=True).start()
-        hides_mock = mock.patch("app.utils.get_hides", spec=True).start()
+        ignored_mock = mock.patch("app.utils.get_ignored", spec=True).start()
 
-        yield cfg_mock, os_walk_mock, get_user_mock, sudoers_mock, hides_mock
+        yield cfg_mock, os_walk_mock, get_user_mock, sudoers_mock, ignored_mock
 
         mock.patch.stopall()
 
-    def test_no_sudores_or_hides(self, folders_mocker):
-        cfg_mock, os_walk_mock, get_user_mock, sudoers_mock, hides_mock = folders_mocker
+    def test_no_sudores_or_ignored(self, folders_mocker):
+        (
+            cfg_mock,
+            os_walk_mock,
+            get_user_mock,
+            sudoers_mock,
+            ignored_mock,
+        ) = folders_mocker
 
         cfg_mock.CLOUD_PATH = "/test"
         os_walk_mock.return_value = self.input_data
         get_user_mock.return_value = None
         sudoers_mock.return_value = []
-        hides_mock.return_value = []
+        ignored_mock.return_value = []
 
         expected = self.output_data[1:]
         real = tuple([x.as_posix() for x in get_folders()])
@@ -183,27 +192,39 @@ class TestGetFolders:
         assert real == expected
 
     def test_sudoers(self, folders_mocker):
-        cfg_mock, os_walk_mock, get_user_mock, sudoers_mock, hides_mock = folders_mocker
+        (
+            cfg_mock,
+            os_walk_mock,
+            get_user_mock,
+            sudoers_mock,
+            ignored_mock,
+        ) = folders_mocker
 
         cfg_mock.CLOUD_PATH = "/test"
         os_walk_mock.return_value = self.input_data
         get_user_mock.return_value = "user"
         sudoers_mock.return_value = ["user"]
-        hides_mock.return_value = []
+        ignored_mock.return_value = []
 
         expected = self.output_data
         real = tuple([x.as_posix() for x in get_folders()])
 
         assert real == expected
 
-    def test_hides(self, folders_mocker):
-        cfg_mock, os_walk_mock, get_user_mock, sudoers_mock, hides_mock = folders_mocker
+    def test_ignored(self, folders_mocker):
+        (
+            cfg_mock,
+            os_walk_mock,
+            get_user_mock,
+            sudoers_mock,
+            ignored_mock,
+        ) = folders_mocker
 
         cfg_mock.CLOUD_PATH = "/test"
         os_walk_mock.return_value = self.input_data
         get_user_mock.return_value = "user"
         sudoers_mock.return_value = []
-        hides_mock.return_value = ["folder-2"]
+        ignored_mock.return_value = ["folder-2"]
 
         expected = self.output_data[1:-2]
         real = tuple([x.as_posix() for x in get_folders()])
@@ -221,3 +242,75 @@ def test_gen_random_password():
 
     p4 = gen_random_password(n=24)
     assert len(p4) == 24
+
+
+@pytest.mark.skipif(sys.platform != "win32", reason="Does not work on linux")
+@mock.patch("app.utils.request", autospec=True)
+class TestGetPostArg:
+    data_req_strip = (
+        ({"data": "value"}, "value"),
+        ({"data": "  value  "}, "value"),
+        ({"data": "  value  \n"}, "value"),
+        ({"data": ""}, RuntimeError),
+        ({"data": "  "}, RuntimeError),
+        ({"data": " \n\n "}, RuntimeError),
+        ({}, RuntimeError),
+    )
+
+    data_not_req_strip = (
+        ({"data": "value"}, "value"),
+        ({"data": "  value  "}, "value"),
+        ({"data": "  value  \n"}, "value"),
+        ({"data": ""}, None),
+        ({"data": "  "}, None),
+        ({"data": " \n\n "}, None),
+        ({}, None),
+    )
+
+    data_req_not_strip = (
+        ({"data": "value"}, "value"),
+        ({"data": "  value  "}, "  value  "),
+        ({"data": "  value  \n"}, "  value  \n"),
+        ({"data": ""}, RuntimeError),
+        ({"data": "  "}, "  "),
+        ({"data": " \n\n "}, " \n\n "),
+        ({}, RuntimeError),
+    )
+
+    data_not_req_not_strip = (
+        ({"data": "value"}, "value"),
+        ({"data": "  value  "}, "  value  "),
+        ({"data": "  value  \n"}, "  value  \n"),
+        ({"data": ""}, ""),
+        ({"data": "  "}, "  "),
+        ({"data": " \n\n "}, " \n\n "),
+        ({}, None),
+    )
+
+    @pytest.mark.parametrize("request_data, expected", data_req_strip)
+    def test_req_strip(self, request_mock, request_data, expected):
+        request_mock.form = request_data
+        if expected == RuntimeError:
+            with pytest.raises(expected):
+                get_post_arg("data", required=True, strip=True)
+        else:
+            assert get_post_arg("data", required=True, strip=True) == expected
+
+    @pytest.mark.parametrize("request_data, expected", data_not_req_strip)
+    def test_not_req_strip(self, request_mock, request_data, expected):
+        request_mock.form = request_data
+        assert get_post_arg("data", required=False, strip=True) == expected
+
+    @pytest.mark.parametrize("request_data, expected", data_req_not_strip)
+    def test_req_not_strip(self, request_mock, request_data, expected):
+        request_mock.form = request_data
+        if expected == RuntimeError:
+            with pytest.raises(expected):
+                get_post_arg("data", required=True, strip=False)
+        else:
+            assert get_post_arg("data", required=True, strip=False) == expected
+
+    @pytest.mark.parametrize("request_data, expected", data_not_req_not_strip)
+    def test_not_req_not_strip(self, request_mock, request_data, expected):
+        request_mock.form = request_data
+        assert get_post_arg("data", required=False, strip=False) == expected